+++ lib/acyort.js
@@ -5,7 +5,7 @@ const markeder = require('./markeder/')
 const Processor = require('./processor/')
 const Render = require('./render/')
 
-const Server = require('./web/')
+const Server = require('./server/')
 
 class Acyort {
   constructor() {
