+++ lib/server/index.js
@@ -1,65 +1,115 @@
 const fs = require('fs-extra')
 const pathFn = require('path')
 const chokidar = require('chokidar')
-const WS = require('ws').Server
-const source = require('../source')
-const acyort = require('../')
-const server = require('./server')
-const { log } = require('../util')
-const config = require('../config')()
+const { Server } = require('ws')
+const StaticServer = require('./server')
+const config = require('../config/')
+const Logger = require('../logger/')
 
-function Index(port) {
-  if (!config.dev && process.env.NODE_ENV !== 'dev' && !config.json) {
-    log.info(`Server running\n=> http://127.0.0.1:${port}/`, '\nCTRL + C to shutdown')
-    return server(port)
+class Ws extends StaticServer {
+  constructor(port) {
+    super(config)
+
+    this.config = config
+    this.port = port
+    this.logger = new Logger()
+
+    this.clients = []
+    this.server = null
+    this.ws = null
+    this.watcher = null
   }
 
-  global.live = true
+  get live() {
+    return this.config.dev && !this.config.json
+  }
 
-  const inject = fs.readFileSync(pathFn.join(__dirname, 'inject.html'), 'utf8')
-  const staticServer = server(port, inject)
-  const WSserver = new WS({ server: staticServer })
-  const watchPath = pathFn.join(process.cwd(), 'themes', config.theme)
-  const watcher = chokidar.watch(watchPath, {
-    ignored: /(^|[/\\])\../,
-    ignoreInitial: true,
-  })
+  log() {
+    const livereload = this.live ? '(LiveReload)' : ''
+    const info = `Server running${livereload}\n=> http://127.0.0.1:${this.port}/`
+    this.logger.info(info, '\nCTRL + C to shutdown')
+  }
 
-  let clients = []
+  initWatcher() {
+    this.watcher = chokidar.watch(this.watchDir, {
+      ignored: /(^|[/\\])\../,
+      ignoreInitial: true,
+    })
+  }
 
-  watcher.on('all', (e, file) => {
-    if (file.indexOf(`${config.theme}/`) > -1) {
-      log.info(e, file)
+  initWs() {
+    this.ws = new Server({ server: this.server })
+  }
+
+  get watchDir() {
+    return pathFn.join(process.cwd(), 'themes', this.config.theme)
+  }
+
+  currentEvent(file) {
+    const { theme } = this.config
+    const source = `${theme}/source`
+    const layout = `${theme}/layout`
+    const i18n = `${theme}/i18n`
+
+    if (file.indexOf(source) > -1) {
+      if (this.fileExt(file) === '.css') {
+        return 'css'
+      }
+      return 'static'
     }
 
-    clients.forEach((client) => {
-      if (file.indexOf(`${config.theme}/source`) > -1) {
-        source()
-        if (pathFn.extname(file) === '.css') {
-          client.send('css')
-        } else {
-          client.send('reload')
+    if (file.indexOf(layout) > -1 && this.fileExt(file) === '.html') {
+      return 'html'
+    }
+
+    if (file.indexOf(i18n) > -1 && this.fileExt(file) === '.yml') {
+      return 'yml'
+    }
+
+    return 'unknow'
+  }
+
+  fileExt(file) {
+    return pathFn.extname(file)
+  }
+
+  _() {
+    this.log()
+    this.server = this.createServer(this.port)
+
+    if (!this.live) {
+      return this.server
+    }
+
+    this.initWatcher()
+    this.initWs()
+
+    this.ws.on('connection', (client) => {
+      client.on('close', () => {
+        this.clients = this.clients.filter(c => c !== client)
+      })
+      this.clients.push(client)
+    })
+
+    this.watcher.on('all', (e, file) => {
+      if (file.indexOf(`${config.theme}/`) > -1) {
+        this.logger.info(e, file)
+      }
+
+      this.clients.forEach((client) => {
+        switch (this.currentEvent(file)) {
+          case 'css':
+            client.send('css')
+            break;
+          case 'static':
+          case 'html':
+          case 'yml':
+            client.send('reload')
+            break;
         }
-      }
-
-      if (file.indexOf(`${config.theme}/layout`) > -1 && pathFn.extname(file) === '.html') {
-        acyort().then(() => client.send('reload'))
-      }
-
-      if (file.indexOf(`${config.theme}/i18n`) > -1 && pathFn.extname(file) === '.yml') {
-        acyort().then(() => client.send('reload'))
-      }
+      })
     })
-  })
-
-  WSserver.on('connection', (ws) => {
-    ws.on('close', () => {
-      clients = clients.filter(client => client !== ws)
-    })
-    clients.push(ws)
-  })
-
-  return log.info(`Server running(LiveReload)\n=> http://127.0.0.1:${port}/`, '\nCTRL + C to shutdown')
+  }
 }
 
-module.exports = Index
+module.exports = Ws
