+++ lib/server/server.js
@@ -1,54 +1,67 @@
 const http = require('http')
 const url = require('url')
-const path = require('path')
+const pathFn = require('path')
 const fs = require('fs-extra')
 const mime = require('mime')
-const config = require('../config/')
 
-function server(port, inject) {
-  const server = http.createServer((req, res) => {
-    let filePath = path.join(process.cwd(), config.public_dir, url.parse(req.url).pathname)
-    filePath = decodeURIComponent(filePath)
+class Server {
+  constructor(config) {
+    this.config = config
+  }
 
-    const page404 = () => {
-      res.writeHead(404, { 'Content-Type': 'text/plain' })
-      res.write('404 Not Found')
-      res.end()
+  getPath(req) {
+    const path = pathFn.join(process.cwd(), this.config.public_dir, url.parse(req.url).pathname)
+    return decodeURIComponent(path)
+  }
+
+  inject(path) {
+    if (
+      !this.config.dev         ||
+      this.config.json  ||
+      mime.lookup(path) !== 'text/html'
+    ) {
+      return ''
     }
 
-    const page500 = (err) => {
-      res.writeHead(500, { 'Content-Type': 'text/plain' })
-      res.write(err)
-      res.end()
-    }
+    return fs.readFileSync(pathFn.join(__dirname, 'inject.html'), 'utf8')
+  }
 
-    if (!fs.existsSync(filePath)) {
-      return page404()
-    }
+  errPage(res, err) {
+    res.writeHead(err ? 500 : 404, { 'Content-Type': 'text/plain' })
+    res.write(err ? err : '404 Not Found')
+    res.end()
+  }
 
-    if (fs.statSync(filePath).isDirectory()) {
-      filePath = path.join(filePath, 'index.html')
-    }
+  createServer(port) {
+    return http.createServer((req, res) => {
+      let path = this.getPath(req)
 
-    if (!fs.existsSync(filePath)) {
-      return page404()
-    }
-
-    return fs.readFile(filePath, 'binary', (err, file) => {
-      if (err) {
-        page500(err)
-      } else {
-        res.writeHead(200, { 'Content-Type': mime.lookup(filePath) })
-        res.write(file, 'binary')
-        if (inject && mime.lookup(filePath) === 'text/html') {
-          res.write(inject)
-        }
-        res.end()
+      if (!fs.existsSync(path)) {
+        return this.errPage(res)
       }
-    })
-  }).listen(port)
 
-  return server
+      if (fs.statSync(path).isDirectory()) {
+        path = pathFn.join(path, 'index.html')
+      }
+
+      if (!fs.existsSync(path)) {
+        return this.errPage(res)
+      }
+
+      return fs.readFile(path, 'binary', (err, file) => {
+        if (err) {
+          return this.errPage(res, err)
+        }
+
+        res.writeHead(200, { 'Content-Type': mime.lookup(path) })
+        res.write(file, 'binary')
+        res.write(this.inject(path))
+        res.end()
+
+        return true
+      })
+    }).listen(port)
+  }
 }
 
-module.exports = server
+module.exports = Server
