--- lib/web/inject.html
@@ -1,37 +1,0 @@
-<script>
-(function() {
-  function socket() {
-    var ws = new WebSocket('ws://'+ location.host)
-    ws.onmessage = function(e) {
-      e = e.data
-      if (e === 'reload') {
-        location.reload()
-      }
-      if (e === 'css') {
-        var sheets = [].slice.call(document.getElementsByTagName('link'))
-        var head = document.getElementsByTagName('head')[0]
-        for (var i = 0; i < sheets.length; i += 1) {
-          var elem = sheets[i]
-          var rel = elem.rel
-          var href = elem.href
-
-          if (rel == 'stylesheet') {
-            head.removeChild(elem)
-            href = href.split('?')[0]
-            elem.href = href + '?_=' + (+new Date())
-            head.appendChild(elem)
-          }
-        }
-      }
-    }
-    ws.onopen = function(e) {
-      console.info('WebSocket connected')
-    }
-    ws.onclose = function(e) {
-      console.error('WebSocket closed, reconnecting...')
-      setTimeout(socket, 5000)
-    }
-  }
-  socket()
-})()
-</script>
